+++ src/mariner/index.ts
@@ -5,4 +5,4 @@ export { Logger, getLogger, setLogger } from '../tab-level-logger';
 export { generateConfluenceMarkup } from '../Utilities/generateConfluenceMarkup';
 export { generateHtml } from '../Utilities/generateHtml';
 export { generateGitHubMarkdown } from '../Utilities/generateGitHubMarkdown';
-export { ContributorFetcher } from '../contributorFetcher';
+export { GitHubContributorFetcher as ContributorFetcher } from '../gitHubContributorFetcher';
